v5.x-remove-runtimecosting: moved InvalidType from special.sigma -> sigmastate.eval
