+++ sigmastate/src/main/scala/sigmastate/eval/Exceptions.scala
@@ -1,4 +1,4 @@
-package special.sigma
+package sigmastate.eval
 
 final class InvalidType(message: String) extends Exception(message)
