+++ sigmastate/src/test/scala/org/ergoplatform/ErgoAddressSpecification.scala
@@ -1,30 +1,30 @@
 package org.ergoplatform
 
-import java.math.BigInteger
-
-import org.ergoplatform.ErgoAddressEncoder.{hash256, MainnetNetworkPrefix, TestnetNetworkPrefix}
+import org.ergoplatform.ErgoAddressEncoder.{MainnetNetworkPrefix, TestnetNetworkPrefix, hash256}
 import org.ergoplatform.SigmaConstants.ScriptCostLimit
 import org.ergoplatform.validation.{ValidationException, ValidationRules}
 import org.scalatest.{Assertion, TryValues}
 import scorex.crypto.hash.Blake2b256
-import sigmastate.basics.DLogProtocol
-import sigmastate.basics.DLogProtocol.{ProveDlog, DLogProverInput}
-import sigmastate.serialization.ErgoTreeSerializer.DefaultSerializer
-import sigmastate.serialization.{ValueSerializer, GroupElementSerializer}
 import scorex.util.encode.Base58
-import sigmastate.{CrossVersionProps, SigmaAnd, SType}
-import sigmastate.Values.{UnparsedErgoTree, Constant, EvaluatedValue, ByteArrayConstant, IntConstant, ErgoTree}
-import sigmastate.eval.IRContext
-import sigmastate.helpers._
+import sigmastate.Values.{ByteArrayConstant, Constant, ErgoTree, IntConstant, UnparsedErgoTree}
+import sigmastate.basics.DLogProtocol
+import sigmastate.basics.DLogProtocol.{DLogProverInput, ProveDlog}
+import sigmastate.eval.{IRContext, InvalidType}
 import sigmastate.helpers.TestingHelpers._
+import sigmastate.helpers._
 import sigmastate.interpreter.ContextExtension.VarBinding
 import sigmastate.interpreter.CryptoConstants.dlogGroup
+import sigmastate.interpreter.Interpreter.{ScriptEnv, ScriptNameProp}
 import sigmastate.interpreter.{ContextExtension, CostedProverResult}
-import sigmastate.interpreter.Interpreter.{ScriptNameProp, ScriptEnv}
 import sigmastate.lang.Terms.ValueOps
-import sigmastate.lang.exceptions.{CosterException, CostLimitException}
+import sigmastate.lang.exceptions.CostLimitException
+import sigmastate.serialization.ErgoTreeSerializer.DefaultSerializer
+import sigmastate.serialization.{GroupElementSerializer, ValueSerializer}
 import sigmastate.utils.Helpers._
-import special.sigma.{InvalidType, SigmaDslTesting}
+import sigmastate.{CrossVersionProps, SType, SigmaAnd}
+import special.sigma.SigmaDslTesting
+
+import java.math.BigInteger
 
 class ErgoAddressSpecification extends SigmaDslTesting
   with TryValues with CrossVersionProps {
