+++ sigmastate/src/test/scala/sigmastate/utxo/BasicOpsSpecification.scala
@@ -13,8 +13,8 @@ import sigmastate.helpers.{ContextEnrichingTestProvingInterpreter, ErgoLikeConte
 import sigmastate.helpers.TestingHelpers._
 import sigmastate.interpreter.Interpreter._
 import sigmastate.lang.Terms._
-import special.sigma.InvalidType
 import SType.AnyOps
+import sigmastate.eval.InvalidType
 import sigmastate.interpreter.ContextExtension.VarBinding
 import sigmastate.interpreter.ErgoTreeEvaluator.DefaultEvalSettings
 import sigmastate.interpreter.{CryptoConstants, EvalSettings}
